requirements.txt
fastapi==0.86.0
uvicorn==0.19.0
pyahocorasick==2.0.0
jieba==0.42.1